Moved cli to __main__.py
